requirements.txt
flask
opencv-python
numpy
face_recognition_models
Pillow
click
colorama